+++ packages/mermaid/src/diagram.spec.ts
@@ -74,13 +74,13 @@ Expecting 'TXT', got 'NEWLINE'"
     const diagram = await getDiagramFromText(`sequenceDiagram
     A->>B: I #9829; you!
     B->>A: I #9829; you #infin; times more!`);
+    //@ts-ignore
     const messages = diagram.db?.getMessages?.();
     if (!messages) {
       throw new Error('Messages not found!');
     }
-    const result = ['I ﬂ°°9829¶ß you!', 'I ﬂ°°9829¶ß you ﬂ°infin¶ß times more!'];
-    messages.forEach((message, index: number) => {
-      expect(message.message).toBe(result[index]);
-    });
+
+    expect(messages[0].message).toBe('I ﬂ°°9829¶ß you!');
+    expect(messages[1].message).toBe('I ﬂ°°9829¶ß you ﬂ°infin¶ß times more!');
   });
 });
